Rename wrap => wrapping
